Use more sophisticated query strategy in delta API.
Summary:
Serially loading objects to be rendered proves to be insufficiently performant,
so instead we undertake the slightly more complicated approach of loading all
objects of a given type at once.

Test Plan:
Unit tests; make same requests against this code and the original
implementation and compare timing and results.

Reviewers: kav-ya

Reviewed By: kav-ya

Differential Revision: https://review.inboxapp.com/D1481