Fix api_client test fixture with non-default test accounts.
Summary:
It's not clear to me that the custom namespace_id substitution ever
worked... the message tests worked using the argument simply because
they happened to also be expecting the default account's namespace,
which is what the api_client fixture was hardcoded to use.

I am fixing this in order to be able to use api_client with an account other
than default_account.

Test Plan: unit tests

Reviewers: emfree

Reviewed By: emfree

Differential Revision: https://review.inboxapp.com/D1485