Tighten error-handling in message parsing.
Summary:
Prior to this patch, we aborted message parsing and marked a message as
broken if *any* MIME part failed to parse. This could cause problems like
having a blank body on a message even though the body MIME part parsed
properly, simply because an attachment part had a filename that broke
the parser.

This patch makes it so we handle errors for each MIME part separately.
We'll still mark the message as having a decode error if any MIME part
is broken, but we will attempt to parse every MIME part regardless of
other MIME parts being broken.

This patch also removes our logging of raw message bodies to disk in the
case of a decode error. Now that we store full message bodies in S3
(regardless of parse errors later), it's simpler and more durable to
access them there than to collect them across all of our sync engine
disks.

Fixes T994.

Test Plan: regression test included

Reviewers: emfree

Reviewed By: emfree

Maniphest Tasks: T994, T944

Differential Revision: https://review.inboxapp.com/D1500