Folder-level tracking of sync_should_run state.
Summary: Part of T1646. In order to easily determine whether an account is in good standing, store the expected run state for a folder directly in that folder's sync status entry. This is somewhat trivial for IMAP (for now) since all folders should be syncing, but will allow us to simplify the state rollup logic regardless.

Test Plan: Added new tests.

Reviewers: emfree

Reviewed By: emfree

Differential Revision: https://phab.nylas.com/D1572