[SCHEMA] The Folders and Labels APIs.
This monstrous commit introduces the folders and labels APIs,
a set of provider-dependent APIs to retrieve and modify the
folders/labels associated with messages and threads:

* Provider-dependent APIs -
Depending on the provider, an account may be organized by folders (Exchange, generic IMAP) or
labels (Gmail); folders and labels have fundamentally different semantics, and these are
preserved in the APIs.

* Expose folders/ labels at both the message, thread -levels -
- The 'message' object returned by the API includes the folder a message is in/ labels associated with it;
the 'thread' object includes a list of all the folders the thread has messages in/ labels on the thread.
- The messages/ threads in a certain folder/ label can be retrieved using a new filter query param ('in')
as well.

* Syncback support to modify the folders/ labels of messages, threads -
A message can be moved between folders/ its labels can be updated;
the thread operations are a convenient shortcut to perform bulk operations on messages,
whereby an operation on a thread is performed on all the messages in the thread.

The folders and labels APIs are intended to eventually replace the tags APIs.
However, both are currently supported and this commit does /not/ remove support
for the latter.

NOTE:
The following changes are also included:
* 'unread', 'starred' as independent message and thread -level attributes
that can be read and modified as above.
* Support for syncing multiple IMAP folders with same flag.