Use a watcher thread to detect ongoing greenlet blocking.
Summary:
Previous instrumentation of mailsync would warn if a greenlet blocked the event
loop, but only after it unblocked. This is not useful if something gets stuck
for a really long time. Instead, use a separate watcher thread to check for
blocking. Consolidate similar code that previously did the same thing for the
API.

Test Plan:
Introduce blocking in both API and mailsync ("while True: pass"),
check that appropriate warning is logged.

Reviewers: kav-ya

Reviewed By: kav-ya

Differential Revision: https://phab.nylas.com/D1615