Redo tests to not depend on dump file.
This allows for substantially faster and more maintainable tests. There are
still many rough edges, but this works. A test database is now generated from
declarative models at the start of a test session instead of being loaded from
a dumpfile.  The `default_account`/`default_namespace` fixtures can be used to
create a new Account/Namespace object for each test function. This means that
for most tests that operate on only a single namespace, any other state in the
test database is irrelevant.