Fix an EAS reconciling bug
Summary:
This bug is funny ;) Some messages (particularly self-sent messages, or sent through a mailing list) were not reconciled correctly. This is because the reconciliation code does this to get all potentially-related messages:

```
 msgs = db_session.query(Message).filter(
        Message.namespace_id == new_message.namespace_id,
        Message.subject == new_message.subject,
        Message.message_id_header == new_message.message_id_header).all()
```

The problem is that when we're sending a draft we're not setting `message_id_header` before sending it, which means we won't find it afterwards.

Test Plan: Tested manually.

Reviewers: kav-ya, emfree

Reviewed By: emfree

Maniphest Tasks: T1904

Differential Revision: https://phab.nylas.com/D1691