A tiny script to make querying objects quicker
Summary:
This script is `get-id`'s little brother. Instead of printing the id of an object, `get-object` fetches the object and drops you inside a REPL with the object in scope.

Gone are the days of writing a one-off script because you need to query a python property of an object!

Test Plan: Tested manually.

Reviewers: spang, jennie

Reviewed By: jennie

Differential Revision: https://phab.nylas.com/D1699