Log bogus iCalendar files and fix this UnboundLocalError
Summary:
I've tried fixing yesterday's UnboundLocalError but without the iCalendar file causing the bug I could only make guesses.

My theory is that somehow we got an event with a `start` as a datetime and an `end` as a date. I've added an assert to check this and logging code to save the iCalendar file. I feel it's better than randomly fixing something and hoping we won't get the same error in two weeks.

Test Plan: Inspected the code.

Reviewers: emfree

Reviewed By: emfree

Subscribers: spang

Projects: #calendars

Maniphest Tasks: T2213

Differential Revision: https://phab.nylas.com/D1712