Fully reset sync status
Summary:
Builds on the 'reset heartbeat when sync starts' logic to also reset the folder sync_should_run flag, and then flip it back to the right state when we know for sure that folder should be running. The goal is to avoid stale state when an account restarts, which can lead to a temporary 'dead' status. There is still a short period between resetting state and having every folder report in, which makes sense and should be dealt with at the reporting level.

Context in T2315.

Test Plan: Restart syncs while observing status.

Reviewers: emfree

Reviewed By: emfree

Differential Revision: https://phab.nylas.com/D1756