Fix PytestTestEngine to work with the latest version of arcanist
Summary:
Updating to the latest version of arcanist broke PytestTestEngine. We were failing with the error `Attempt to write to undeclared property`.

It seems that the phab developers changed a couple properties on the object and it's now read-only. Fixed this by using a local variable.

Test Plan: The tests run.

Reviewers: jennie

Reviewed By: jennie

Differential Revision: https://phab.nylas.com/D1777