Messages sent via our api now added to sent folder after sending.
Summary:
Messages sent via our API are not assigned categories until they are synced with the provider. This causes issues when
the message is passed to thread.update_from_message. The sent message has not been assigned the sent category, so the
receivedrecentdate attribute on the thread is updated with every message, instead of messages that are not "sent"

Test Plan: Unit test. Tested manually

Reviewers: kav-ya

Subscribers: emfree

Differential Revision: https://phab.nylas.com/D1811