Depend on nylas-production-python.
Summary:
nylas-production-python is new library which lives at
https://github.com/nylas/nylas-production-python and contains two pieces split
out from the sync engine:

1. our logging configuration, including sentry integration
2. our special WSGI adaptor for running in production

It's useful to split this out into its own library so we can use it in all
of our production server-side Python code without pulling in the entire sync
engine as a dependency.

We rarely change this code anymore, so it shouldn't be a hassle to have it
in its own repo. To change code in the library, we require a version bump and
a release to PyPI.

I've already shipped version 0.1.0 to PyPI. Only thing left to do is update
the repo's README to be a bit more illuminating on its use and maintenance.

Test Plan: unit tests, run sync locally

Reviewers: emfree

Reviewed By: emfree

Differential Revision: https://phab.nylas.com/D1866