Save IMAP UIDNEXT value.
Summary:
We'll soon be able to use this for much improved generic IMAP sync. In order to
be able to efficiently migrate, start saving the reported UIDNEXT value
whenever we poll a folder. (We're not actually using it for anything yet.)

Test Plan: run syncs, verify that uidnext values are persisted.

Reviewers: spang

Reviewed By: spang

Differential Revision: https://phab.nylas.com/D1892