Allow accounts with 'killed' state to be considered deleted.
Summary:
Not totally sure this is exactly the right fix here. It seems that
account.disable_sync() doesn't change account state to 'stopped'--so a
killed account which is marked as deleted will falsely trigger the 'make
sure account is marked as deleted' check at the beginning of
bin/delete-account-data, preventing account deletion.

The other possible fix would be to mark accounts as 'stopped' in
disable_sync(). If that is preferable, let me know. I chose this
strategy because I'm not sure if we want to preserve the 'killed' state
for any reason.

Test Plan: unit tests, manual testing

Reviewers: emfree

Reviewed By: emfree

Subscribers: kav-ya

Differential Revision: https://phab.nylas.com/D1952