Normalize "regex" patterns in providers.py.
Summary:
The immediate reason to do this is because the current code is wrongly
interpreting 'ppsmtp.com' to be an MX server for Gmail, but altogether
the current plugging of things that were intended to be straight strings
into a regex has lots of room for misinterpretation, mostly with the '.'
character, which can stand for everything—but we generally intend for it
to be used literally, except in wildcard matches.

So as a compromise between legibility and being able to actually tell
what a given pattern does, we leave dots as literal dots and * as a glob
*, and substitute these into proper regexes before using them. Otherwise
we'd have to switch out '.' for '[.]' and '*' for '.*' everywhere.

Test Plan:
Unit tests, some manual testing

I added some new test cases for providers, but couldn't find good domains for
namecheap/godaddy that aren't likely to randomly stop working at some point
in the future.

Reviewers: khamidou

Reviewed By: khamidou

Differential Revision: https://phab.nylas.com/D2268