Suppress database connection pooling behavior in sync scheduling.
Let's say you have lots of sync engines and lots of MySQL shards. And say you
have de facto affinity of sync engines to shards. That is, a sync
engine is only actually responsible for accounts on one shard in general.
However, the sync engine still checks /all/ shards for account syncs to run,
and as a result acquires a persistent connection to each. MySQL doesn't like
having lots of connections. So for now at least, just explicitly close the
connection.

Of course, this is just a standin until a more fully-developed solution for
large-scale sync scheduling is fixed.