Include the account provider in Sentry exceptions.
Summary:
To correlate sync errors with providers.
We introduce the Account.verbose_provider property to expose the specific provider (Office365 versus Outlook etc.)
since this is what we care about. Ideally we'd directly use the existing Account.provider property,
but the existing (and hella messy) auth code uses this in all sorts of ways - for example, to determine the
auth_handler to use and the provider icon to display on the signup page.
So for now, in order to get at these crucial metrics in a timely manner, we use a new property instead
rather than embarking on a giant, cross-repo refactor.

Test Plan:
Unit test added.
Gmail/ Exchange/ Yahoo syncs run locally.

Reviewers: khamidou

Reviewed By: khamidou

Differential Revision: https://phab.nylas.com/D2312