Move bitrotted Docker config to https://github.com/nylas/sync-engine-docker
We don't actually use Docker to run anything in production, so this code is not consistently tested or supported. New developers are confused when they check out the project and assume that we support Docker, so I'm removing it.

I've created a new repo at https://github.com/nylas/sync-engine-docker for development of a Docker wrapper to run the sync engine. I think this is a better home for it, since the sync engine can be deployed a bunch of different ways.