Introduce a separate test config for production builds.
Summary:
The Jenkins job sets the INBOX_CFG_PATH to point to this config
when running tests against the production build.

This is the simplest way to accomplish what we want
namely have concurrently running master/ production builds
complete without hanging due to tests contending for the same database
tables.

Other options considered were having a single test config
but selectively processing it/ updating it on the fly.

The corresponding changes to the Jenkins job definitions can be found here:
https://github.com/nylas/jenkins-job-definitions/commit/fc900f729fcc83f7be69b51158486efabe952808

Test Plan:
Tested manually.
- Unit tests pass with/ without setting the INBOX_CFG_PATH
- Dropped all test databases, revoked user permissions,
ran bin/create-test-db and then the tests as above.

Reviewers: emfree

Reviewed By: emfree

Differential Revision: https://phab.nylas.com/D2396