Rework account scheduling (part 1): Allow process specification.
Summary:
Allow accounts to be scheduled to run on a particular process, rather than
merely on a particular host. This is done by specifying
`account.sync_host = "<hostname>:<process number>"`.

For the moment, we support both this, as well as the prior method for
specifying an account's sync host. This way we can transition smoothly or roll
back easily.

Also remove host-local account "sync locking". This is essentially useless, but
can break sync if an account is unscheduled, then rescheduled on the same
process.

Test Plan: Unit tests updated.

Reviewers: spang

Reviewed By: spang

Differential Revision: https://phab.nylas.com/D2244