Allow shared folders to work when VAGRANT_CWD is used
Summary:
Previously when VAGRANT_CWD was used and the current working directory
wasn't the sync-engine checkout, the synced folders would fail to be
found by "vagrant reload/up" because of the relative path used to find
them. This has been fixed by using a glob relative to the directory
which contains the Vagrantfile.

Also:
- use a more precise glob pattern to avoid needing to filter
  directories in the loop
- confirm that each "share-*" entry is actually a directory

Test Plan: Manually tested with "vagrant reload/up"

Reviewers: emfree, kav-ya

Reviewed By: kav-ya

Differential Revision: https://phab.nylas.com/D2430