Bypass database loads on unchanged IMAP flags refresh.
Flag refreshes for generic IMAP folders still account for a depressingly large
fraction of overall workload. But some optimization is possible. If a flags
fetch response is exactly the same as the last one we got from the server, then
there are no changes to commit, and so we don't need to issue any (potentially
expensive) database loads.