Replace sync-host-to-shard affinity with 'zone' configuration.
Summary:
We no longer need or want to impose the constraint that "accounts on shard 1
should only be synced by host A". But it is useful to say "accounts on shards
in us-west-2b should only be synced by hosts in us-west-2b", for example. This
patch introduces configuration support for that. Specifically, read a 'ZONE'
value from the configuration of a sync host. If configured, that host will only
claim account syncs from shards that reside in the same zone.

Test Plan: Unit tests updated.

Reviewers: kav-ya

Reviewed By: kav-ya

Differential Revision: https://phab.nylas.com/D2429