A script to backfix IMAP UTF-7 labels
Summary:
Because of a bug in IMAPClient, we have 400 production accounts with duplicate, badly-encoded labels (https://github.com/nylas/sync-engine/issues/212).
The bug has since been fixed, but the badly-encoded labels still show up even though they've been marked as deleted because the sync-engine still sees messages with those labels.

This script fixes this by going through the list of accounts, finding such labels and refreshing the flags for those. The messages then get affected to the correct labels, which cause the bad labels to be GCed.

Test Plan: Tested locally and on staging.

Reviewers: emfree

Reviewed By: emfree

Differential Revision: https://phab.nylas.com/D2484