[SCHEMA] Introduce AccountTransaction table.
Summary:
Fixes T5980.

Account object events are recorded in this table in addition to the Transaction table.

The former is to maintain a total ordering over all events for an account;
the latter is to simplify the expensive query by the Webhook service.

NOTE:
I considered writing the back-population script such that AccountTransaction records
are created for existing accounts linked to a webhook-enabled app only to begin with,
but do not think it's worth the complexity it might add later (-- "why are these accounts
missing AccountTransaction records"?).

Test Plan: Tests added.

Reviewers: emfree

Reviewed By: emfree

Maniphest Tasks: T6617, T5980

Differential Revision: https://phab.nylas.com/D2528