Add metadata store
Summary:
Adds a metadata store capable of storing arbitrary JSON against
API objects.

Includes a new database table for Metadata and some minor changes
to accommodate it. Alters the delta API to exclude metadata by
default, and uses a new param `exclude_metadata=False` to include it.
The API for all other access to metadata will live in Redwood.

Note that Metadata entries will remain even if the API object they are
tied to is deleted.

Also note, the new Metadata table contains columns with data about
`Application` objects, which reside in a different database and are
defined outside this repo, in Redwood.

D2468 contains tests for this diff and requires it to run.

Test Plan: New tests in redwood

Reviewers: kav-ya, emfree

Reviewed By: kav-ya, emfree

Subscribers: bengotow

Differential Revision: https://phab.nylas.com/D2467